merge

+++ src/lua_script/lua_script_manager.cpp
@@ -98,7 +98,7 @@ void LuaScript::loaded(FS::IFile& file, bool success, FS::FileSystem& fs)
 	if (success)
 	{
 		m_properties.clear();
-		m_source_code.set((const char*)file.getBuffer(), file.size());
+		m_source_code.set((const char*)file.getBuffer(), (int)file.size());
 		parseProperties();
 		m_size = file.size();
 		decrementDepCount();
